Add various skins to test against with attribution
