+++ valhalla/tests/assets/README.md
@@ -1,0 +1,18 @@
+# Image credits
+
+## Directory: Bad
+
+| Image | Attribution
+| --- | ---
+| `bad_format.jpg` | Alberto Ghizzi Panizza, via [Dreamstime](https://www.dreamstime.com/aghizzi_info)
+| `bad_size.png` | Chimera
+
+## Directory: Good
+| Image | Attribution
+| --- | ---
+| `64x64.png` | [Mojang](https://www.minecraft.net)
+| `128x128.png` |  Chimera
+| `256x256.png` | Chimera
+| `512x256.png` | [Idze](https://www.deviantart.com/idze)
+| `1024x512.png` | [Idze](https://www.deviantart.com/idze)
+| `1024x1024.png` | Chimera
