+++ valhalla/tests/assets/bad/bad_image.txt
@@ -1,0 +1,1 @@
+This is a bad image. In fact, you might say it isn't an image at all!
