+++ valhalla/tests/assets/good/1024x1024.txt
@@ -1,0 +1,1 @@
+34bba71116779d76f31aadada3ed8c257449f066
