+++ valhalla/tests/assets/good/128x128.txt
@@ -1,0 +1,1 @@
+1d94ec231fe293ecd80391b24b7de71149f2e889
