+++ valhalla/tests/assets/good/256x256.txt
@@ -1,0 +1,1 @@
+f1d8d2725d2e3c918d796cdfbb6f274f47eb78b4
