+++ valhalla/tests/assets/good/64x64.png
