+++ valhalla/tests/assets/good/64x64.txt
@@ -1,0 +1,1 @@
+18b1728ad94f13571d903b913f5e58d47cee1b0a
