+++ valhalla/tests/test_legacy.py
@@ -4,7 +4,7 @@ from uuid import uuid4
 import pytest
 
 from .conftest import TestClient, TestUser
-from .test_textures import steve_file, steve_hash, steve_url
+from .test_app import steve_file, steve_hash, steve_url
 
 
 def build_request_kwargs(file: str | Path):
